% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHGNC.R
\name{addHGNC}
\alias{addHGNC}
\title{Adding hg19's attributes as new columns to dataframe}
\usage{
addHGNC(df, coords = c("chromosome", "position"))
}
\arguments{
\item{df}{A dataframe with genomic coordinates as columns.}

\item{coords}{A character vector indicating the column names of corresponding
and required genomic coordinates.}
}
\value{
A dataframe with new columns.
}
\description{
`addHGNC()` takes a dataframe and adds matching HGNC symbols, gene biotype
and gene description columns to it.
}
