% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToGRange.R
\name{convertToGRange}
\alias{convertToGRange}
\title{Converting data to Genomic Range}
\usage{
convertToGRange(df, cnames = c("chromosome", "position"))
}
\arguments{
\item{df}{A dataframe as returned by findAtPhenotypes().}

\item{cnames}{A character vector with either two or three strings, each
indicating the corresponding column name for chromosome, start and end
positions. If no end position column is present, indicate only first two
strings and we'll use "start + 1" for end position.}
}
\value{
A GRange object.
}
\description{
`convertToGRange()` converts a dataframe with search results into a
genomic range with metadata columns.
}
\examples{
\dontrun{convertToGRange(findAtPhenotypes(dfgroup, "pval", 1e-5))}

}
