% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defGroup.R
\name{defGroup}
\alias{defGroup}
\title{Downloading phenotype files}
\usage{
defGroup(group, phenotypes, sexes, index = TRUE, clean = FALSE)
}
\arguments{
\item{group}{A character string naming all phenotypes. A sub-folder will be
created (as needed) at the ukbbgwas.path download directory.}

\item{phenotypes}{A character vector with phenotype codes as they appear in
neale.manifest$Phenotype.Code.}

\item{sexes}{A character vector of the same length as `phenotypes`, with a
1:1 correspondence specifying one of each sex using the corresponding
character string: "male", "female" or "both_sexes".}

\item{index}{A boolean value indicating if an "index.tsv" file with
phenotype's code and descriptions should be placed inside download
folder.}
}
\value{
None
}
\description{
defGroup() downloads specified phenotype files into a chosen directory named
as a group. This folder will reside inside ~/.ukbbgwas/ by default. Select
the download directory by issuing: options(ukbbgwas.path = "/path/new") or
soft-linking the new path to the default location using your shell:
"$ ln -s /path/new ~/.ukbbgwas".
}
\details{
If the directory specified at "ukbbgwas.path" doesn't exists or is empty,
this function will create it for downloading first 9 core files that are
essential. They are listed as the first 9 rows of the manifest, see
data(neale.manifest)[1:9, ]. Most of these files are described as answers in
the FAQ: <https://www.nealelab.is/uk-biobank/faq>. Please take a look at
'README' file in the `ukbbgwas.path` specified location.

After running this function a "neale.manifest" dataframe will appear in the
Global Environment if not already present.

Defining groups with too many files risks using all your available random
acces memory (RAM) when running searches with findAtPhenotypes() or further
on downstream analysis. In such cases your system will seem unresponsive or
either it will halt entirely. Avoid exceeding your limits by splitting groups
intelligently. As a rule of thumb, consider that every 2-3 files uses 1 Gb of
available RAM later on.
}
\examples{
\dontrun{defGroup("Gestational", c("O13", "O14"), rep("female", 2))}
\dontrun{defGroup("Posdoc",
                  c(c("O13", "O14", "4041", "HYPTENSPREG", "3849"),
                    c("IV_ENDOCRIN_NUTRIT", "23105_irnt", "E66", "E4_OBESITY",
                      "I9_HYPTENS", "I10")),
                  c(rep("female", 5),
                    rep("both_sexes", 6)))}

}
