% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAtPhenotypes.R
\name{findAtPhenotypes}
\alias{findAtPhenotypes}
\title{Searching in all phenotypes from a given group.}
\usage{
findAtPhenotypes(
  df.group,
  search.mode,
  search.args,
  p.threshold = NA,
  write = FALSE,
  lowconfilter = FALSE
)
}
\arguments{
\item{df.group}{A dataframe with all the phenotypes in a given group, just
like the one returned from `lsGroup()`.}

\item{search.mode}{A string keyword enabling one of the available modes,
either "pval", "coords" or "region".}

\item{search.args}{A number to use as maximun in case of "pval" mode. Or a
vector with coordinates in the format `chr:position:ref:alt` (i.e.
`c("X:349813:T:C", "3:843028:A:G")`) in case of "coords" mode. Last, but
not least, the "region" mode would take coordinates too but only for a
single genomic interval using a vector of three integers, namely:
chromosome, start and end position.}

\item{p.threshold}{Optionally, filter out matching hits with p-values above a
given number (i.e. > 5e-10). This argument only makes sense when mode is
not "pval".}

\item{write}{Optionally, set to TRUE and write to a file in "ukbbgwas.path"
(package option) instead of returning a dataframe (that must fit in RAM).
The filename will start with a timestamp (year, month, day, hour, minute)
followed by "_search_", the corresponding phenotype group, search.mode in
use and ending with the ".tsv" extension.}

\item{lowconfilter}{Optionally, set to TRUE to filter out variants that
are low confidence under the heuristic(s) applied by Neale's lab.
(Case/control phenotypes: expected_case_minor_AC < 25 or minor_AF <
0.001. Categorical phenotypes with less than 5 categories:
expected_min_category_minor_AC < 25 or minor_AF < 0.001 Quantitative
phenotypes: minor_AF < 0.001.)}
}
\value{
A dataframe containing variants that match the search criteria.
}
\description{
`findAtPhenotypes()` is this package main function. It searches through all
phenotypes in a previously defined group. There are three search modes, either
by a threshold p-value, by a given vector of coordinates in the format
`chr:position` or by a region from any chromosome.
}
\examples{
\dontrun{findAtPhenotypes(lsGroup("MyResearch"), "pval", 5e-20)}
findAtPhenotypes(lsGroup("example"), "pval", 5e-3)
findAtPhenotypes(lsGroup("example"), "coords", "1:69487:G:A")
\dontrun{findAtPhenotypes(lsGroup("MyResearch"), "coords", with(getVariants("rs236918"), variant))}
findAtPhenotypes(lsGroup("example"), "region", c(1, 69200, 69700))
findAtPhenotypes(lsGroup("example"), "region", c(1, 69200, 69700), 5e-10)

}
