% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHG19.R
\name{generateM37}
\alias{generateM37}
\title{Generate a hg19 table with biomaRt attributes}
\usage{
generateM37()
}
\arguments{
\item{columntributes}{A character vector indicating biomaRt attributes to
retrieve. See:
<https://www.ensembl.org/info/data/biomart/biomart_r_package.html> for
help.}
}
\value{
A dataframe with each of the 'columntributes' and corresponding
    genomic coordinates.
}
\description{
`generateHG19()` retrieves metadata columns using biomaRt.
}
\details{
User won't need this function because we already provide the returned object
built-in the package. It only needs to run `data(hg19)`.
}
