% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenoDescrip.R
\name{getPhenoDescrip}
\alias{getPhenoDescrip}
\title{Retrieving Phenotype Description}
\usage{
getPhenoDescrip(pheno, sex)
}
\arguments{
\item{pheno}{Phenotype code factor.}

\item{sex}{Sex factor.}
}
\value{
A character vector of length 3 with sex, phenotype.code and phenotype
    description; ready to use as tittle in a manhattan plot.
}
\description{
`getPhenoDescrip()` retrieves phenotype description for a paired phenotype
and sex.
}
