% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypePath.R
\name{getPhenotypePath}
\alias{getPhenotypePath}
\title{Retrieving phenotype file path.}
\usage{
getPhenotypePath(phenotypeName, pairedSex, groupName)
}
\arguments{
\item{phenotypeName}{A string with the phenotype name.}

\item{pairedSex}{A string with the corresponding sex to the specified
phenotype.}

\item{groupName}{A string indicating the previously defined group where
phenotype was chosen and downloaded.}
}
\value{
A string with the absolute path to desired phenotype file.
}
\description{
`getPhenotypePath()` retrieves the phenotype file path with its full name and
the previously defined group sub-folder.
}
\examples{
\dontrun{getPhenotypePath("Z85", "both_sexes", "MyResearch")}
}
