% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariants.R
\name{getVariants}
\alias{getVariants}
\title{Searching through variants files.}
\usage{
getVariants(vectorstr)
}
\arguments{
\item{vectorstr}{A character vector corresponding to SNP rs identification
number(s) to retrieve.}
}
\value{
Data frame with variant information. All columns are inherited from
    Neale's Lab published file (compressed TSV format).
}
\description{
`getVariants()` looks for matching variants.
}
\examples{
\dontrun{getVariants(c("rs236918", "rs2369"))}

}
