% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGroup.R
\name{lsGroup}
\alias{lsGroup}
\title{Listing downloaded files.}
\usage{
lsGroup(groupName)
}
\arguments{
\item{groupName}{one of the previously defined groups with `defGroup()`.}
}
\value{
A dataframe with each of the phenotype-sex pairs available in the
    group.
}
\description{
`lsGroup()` lists downloaded files in a corresponding group.
}
\examples{
\dontrun{lsGroup("MyResearch")}
lsGroup("example")

}
