% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{plotRegion}
\alias{plotRegion}
\title{Plotting variants in tracklayer style.}
\usage{
plotRegion(data, start, end, writeFiles = "ALL", fname = NA, group = NA)
}
\arguments{
\item{data}{A dataframe as returned by `findAtPhenotypes(mode = "pval")`.}

\item{start}{An integer value indicating where the loci and plot should
start.}

\item{end}{An integer value indicating where the loci and plot should end.}

\item{writeFiles}{One of "PNG", "PDF", "RDATA", "ALL" or "NONE" indicating if
you want the files to be written to disk or not; and in which format.}

\item{fname}{A character string to name file(s) to write.}

\item{group}{Character string corresponding to group name where
findAtPhenotypes was run to generate the data. This will be used as a
subfolder inside 'ukbbgwas.path' for saving images.}
}
\value{
A Gviz plot object, if saved to a variable; retrieving with print()
    is unsupported by upstream. Only when "writeFiles" option is set to
    "none".
}
\description{
`plotRegion()` plots `-log10(p-values)` along genomic coordinates, as a track
of a genome browser. Ideogram with chromosome position and track with
transcript RNA are added.
}
\details{
PNG & PDF files will be saved in path specified by "ukbbgwas.path" option.
}
\examples{
\dontrun{plotRegion(myFoundByPval, gene_start - 1000, gene_end + 1000, fname = "myLoci")}
\dontrun{p <- plotRegion(myFoundByPval, gene_start, gene_end, "none")}

}
