% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariants.R
\name{plotVariants}
\alias{plotVariants}
\title{Plotting variants in skyscrapper style.}
\usage{
plotVariants(
  data,
  group = NA,
  writeFiles = "ALL",
  color = c("#FF7F0EFF", "#2CA02CFF"),
  cutoff.color = "purple",
  cutoff = NULL,
  writeLoci = FALSE
)
}
\arguments{
\item{data}{A dataframe as returned by `findAtPhenotypes(mode = "pval")`.}

\item{group}{Character string corresponding to group name where
findAtPhenotypes was run to generate the data. This will be used as a
subfolder inside 'ukbbgwas.path' for saving images.}

\item{writeFiles}{One of "PNG", "PDF", "RDATA", "ALL" or "NONE" indicating if
you want the files to be written to disk or not; and in which format.}

\item{color}{A character vector with colors to pass into
`ggbio::plotGrandLinear()` function.}

\item{cutoff.color}{A color to use in cutoff line and value.}

\item{cutoff}{A numeric vector with cutoff values for each pair of phenotypes
- sexes. Pairs are ordered alphabetically. If this vector is shorter of
larger than the number of pairs, an heuristic cutoff value will be used
in each plot. In such case, writeLoci will be disabled.}

\item{writeLoci}{A boolean value indicating if you want a TSV file
accompanying each plot with the corresponding HGNC symbol and chromosome
position of variants above cutoff value. Beware choosing low cutoff
values because too many positions may slow the procedure considerably. If
no group name was passed as argument and/ or writeFiles is "none"; these
TSV files will be written in "ukbbgwas.path" root directory.}
}
\value{
A vector of figures you may retrieve with print(). Only when
    "writeFiles" option is set to "none".
}
\description{
`plotVariants()` plots `-log10(p-values)` along genomic coordinates. This
kind of plot is named manhattan, grandlinear or skyscrapper.
}
\details{
PNG & PDF files will be saved in path specified by "ukbbgwas.path" option.
Each file corresponds to a pair of Phenotype and Sex (as defined when
creating the group).

Heuristic for cutoff is six times the mean `-log10(p-values)`. This cutoff is
finally rounded so that it has no decimal positions.
}
\examples{
\dontrun{plotVariants(data, group = "example", writeFiles = "all", cutoff = TRUE)  # will trigger heuristic}
\dontrun{plots <- plotVariants(data, writeFiles = "none"); print(plots[[1]])}

}
