% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVariantColumn.R
\name{splitVariantColumn}
\alias{splitVariantColumn}
\title{Splitting variant column in data}
\usage{
splitVariantColumn(df)
}
\arguments{
\item{df}{A dataframe as returned by findAtPhenotypes()}
}
\value{
A dataframe with variant column removed and four new columns added.
}
\description{
`splitVariantColumn()` splits variant column into "chromosome", "position",
"ref" and "alt" columns.
}
\details{
This functions is auxilliary for convertToGRange().
}
